+++ markhub/services/github_repository.py
@@ -1,14 +1,18 @@
-from pathlib import Path
-from typing import Dict, List, Union, Optional
+from datetime import datetime
+from pathlib import Path, PurePosixPath
+from typing import Dict, List, Optional, Union
 
 from django.contrib.auth.models import User
+from django.core.exceptions import PermissionDenied
 from django.http import Http404
 from django.http.request import HttpRequest
-
+from django.utils.html import format_html
 from github import Github, UnknownObjectException
+from github.ContentFile import ContentFile
 from github.Repository import Repository
+from markhub.models import PrivatePublish
+from markhub.settings import log_error_with_404, logger
 
-from markhub.settings import logger
 
 @logger.catch
 def get_github_handler(user: User) -> Union[Github, None]:
@@ -53,27 +57,123 @@ class GitHubRepository:
                         self.save_current_branch(request, self.handler.default_branch)
                         logger.info(f"{user.username}/{repo_name} have got from GitHub")
                 except UnknownObjectException as e:
-                    logger.error(f"Repository not found - {e}")
-                    raise Http404(f"Repository not found - {e}")
+                    log_error_with_404(f"Repository not found - {e}")
         if self.handler:
+            self.username = self.handler.owner.login or request.user.username
             request.session['__current_repo__'] = repo_name
 
-    def save_current_branch(self, request: HttpRequest, branch: str) -> None:
-        """ Save repository in session
-        
-        Args:
-            request: Django request object
-            branch: branch name
-        """
-        self.branch = branch
-        request.session[f'{self.handler.name}__current_branch'] = self.branch
+    def create_file(self, path: str, content: str, branch: str = '') -> str:
+        """Create a new file in the repository if success otherwise raise 404 exception
 
-    @property
-    def name(self) -> Optional[str]:
-        """Returns repository name"""
-        if self.handler:
-            return self.handler.name
+        Args:
+            path (str): path to the new file
+            updated_content (str): new file content
+            branch (str): repository branch. Defaults to '' (current repository branch)
+
+        Returns:
+            str: success message in html
+        """
+        branch = branch if branch else self.branch
+        try: 
+            status: dict = self.handler.create_file(
+                path=path, 
+                message=f"Add {PurePosixPath(path).name} at MarkHub", 
+                content=content, 
+                branch=branch
+            )
+            return format_html(
+                'File {} was successfully created with commit <a href="{}" target="_blank">{}</a>.',
+                path,
+                status["commit"].html_url,
+                status["commit"].sha[:7]
+            )
+        except UnknownObjectException as e:
+            log_error_with_404(f"File not created - {e}")
     
+    def delete_file(self, path: str, branch: str = '') -> str:
+        """Delete a file in the repository if success otherwise raise 404 exception
+
+        Args:
+            path (str): path to the deleted file
+            branch (str): repository branch. Defaults to '' (current repository branch)
+
+        Returns:
+            str: success message in html
+        """
+        branch = branch if branch else self.branch
+        try:
+            contents = self.handler.get_contents(path, ref=branch)
+            status: dict = self.handler.delete_file(
+                contents.path, 
+                f"Delete {PurePosixPath(path).name} at MarkHub", 
+                contents.sha, 
+                branch
+            )
+            return format_html(
+                    'File {} was successfully deleted with commit <a href="{}" target="_blank">{}</a>.',
+                    path,
+                    status["commit"].html_url,
+                    status["commit"].sha[:7]
+            )
+        except UnknownObjectException as e:
+            log_error_with_404(f"Path not found - {e}")
+
+    def get_contents(self, path: str, branch: str) -> ContentFile:
+        """Get contents for path, otherwise raise Http404 exception
+
+        Args:
+            path (str): repository item path
+            branch (str): repository branch
+
+        Returns:
+            ContentFile: repository item contents
+        """
+        if not branch:
+            branch = self.branch
+        try:
+            return self.handler.get_contents(path, ref=branch)
+        except UnknownObjectException as e:
+            log_error_with_404(f"Path not found - {e}")
+    
+    def get_context(self, path: str, extra: Dict) -> Dict:
+        """Get template context dict with repository data
+
+        Args:
+            path: repository item path
+            extra: dictionary to extend result
+        
+        Returns:
+            Dict: template context
+        """
+        context = {
+            'username': self.username,
+            'repo': self.name,
+            'private': self.handler.private,
+            'branch': self.branch,
+            'branches': self.branches,
+            'path': path,
+        }
+        if path:
+            context['path_parts'] = self.get_path_parts(path)
+            context['parent_path'] = str(Path(path).parent)
+        context.update(extra)
+        context['history_url'] = 'https://github.com/{username}/{repo}/commits/{branch}/{path}'.format(**context)
+        return context
+
+    def get_file_last_update(self, path: str, branch: str) -> Optional[datetime]:
+        """Get file last update in the branch
+
+        Args:
+            path (str): _file path_
+            branch (str): _repo branch_
+
+        Returns:
+            datetime: _file last update_ or None
+        """
+        commits = self.handler.get_commits(sha=branch, path=path)
+        if commits.totalCount:
+            return commits[0].commit.committer.date
+
     def get_path_parts(self, path: str) -> Dict:
         """ Get path parts dict for path
         
@@ -89,3 +189,69 @@ class GitHubRepository:
         for i in range(len(path_parts)):
             path_parts_dict[path_parts[i]] = '/'.join(path_parts[:i+1])
         return path_parts_dict
+    
+    @property
+    def name(self) -> Optional[str]:
+        """Returns repository name"""
+        if self.handler:
+            return self.handler.name
+    
+    def save_current_branch(self, request: HttpRequest, branch: str) -> None:
+        """ Save repository in session
+        
+        Args:
+            request: Django request object
+            branch: branch name
+        """
+        self.branch = branch
+        request.session[f'{self.handler.name}__current_branch'] = self.branch
+    
+    def update_file(self, path: str, updated_content: str, branch: str = '') -> str:
+        """Update a file in the repository if success otherwise raise 404 exception
+
+        Args:
+            path (str): path to the updated file
+            updated_content (str): updated content
+            branch (str): repository branch. Defaults to '' (current repository branch)
+
+        Returns:
+            str: success message in html
+        """
+        branch = branch if branch else self.branch
+        try:
+            contents = self.get_contents(path, branch)
+            status: dict = self.handler.update_file(
+                path=path, 
+                message=f"Update {PurePosixPath(path).name} at MarkHub", 
+                content=updated_content,
+                sha=contents.sha,
+                branch=branch)
+            return format_html(
+                'File {} was successfully updated with commit <a href="{}" target="_blank">{}</a>.',
+                path,
+                status["commit"].html_url,
+                status["commit"].sha[:7]
+            )
+        except UnknownObjectException as e:
+            log_error_with_404(f"File not updated - {e}")
+
+
+def get_repository_or_error(request: HttpRequest, repo: str) -> GitHubRepository:
+    """Get GitHubRepository instance or raise 404
+
+    Args:
+        request (HttpRequest): Django request instance
+        repo (str): Repository name
+
+    Raises:
+        Http404: if GitHubRepository is not created
+        PermissionDenied: if user has no permission to work with private repositories
+
+    Returns:
+        GitHubRepository: GitHubRepository instance
+    """
+    if repository := GitHubRepository(request, repo): 
+        if repository.handler.private and not request.user.has_perm('markhub.private_repos'):
+            raise PermissionDenied
+        return repository
+    raise Http404("Repository not found")